+++ src/mol-model/structure/structure/structure.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2017-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2017-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
@@ -23,12 +23,10 @@ import { Carbohydrates } from './carbohydrates/data';
 import { computeCarbohydrates } from './carbohydrates/compute';
 import { Vec3, Mat4 } from '../../../mol-math/linear-algebra';
 import { idFactory } from '../../../mol-util/id-factory';
-import { GridLookup3D } from '../../../mol-math/geometry';
 import { UUID } from '../../../mol-util';
 import { CustomProperties } from '../../custom-property';
-import { AtomicHierarchy } from '../model/properties/atomic';
 import { StructureSelection } from '../query/selection';
-import { getBoundary, Boundary } from '../../../mol-math/geometry/boundary';
+import { Boundary } from '../../../mol-math/geometry/boundary';
 import { ElementSymbol } from '../model/types';
 import { CustomStructureProperty } from '../../../mol-model-props/common/custom-structure-property';
 import { Trajectory } from '../trajectory';
@@ -788,8 +786,6 @@ namespace Structure {
         return create(units, { representativeModel: first!, label: first!.label });
     }
 
-    const PARTITION = false;
-
     /**
      * Construct a Structure from a model.
      *
@@ -807,20 +803,18 @@ namespace Structure {
             const operator = atomicChainOperatorMappinng.get(c) || SymmetryOperator.Default;
             const start = chains.offsets[c];
 
-            // set to true for chains that consist of "single atom residues",
-            // note that it assumes there are no "zero atom residues"
-            let singleAtomResidues = AtomicHierarchy.chainResidueCount(model.atomicHierarchy, c) === chains.offsets[c + 1] - chains.offsets[c];
-
-            let multiChain = false;
+            let isMultiChain = false;
+            let isWater = false;
 
             if (isWaterChain(model, c)) {
+                isWater = true;
                 // merge consecutive water chains
                 while (c + 1 < chains.count && isWaterChain(model, c + 1 as ChainIndex)) {
                     const op1 = atomicChainOperatorMappinng.get(c);
                     const op2 = atomicChainOperatorMappinng.get(c + 1 as ChainIndex);
                     if (op1 !== op2) break;
 
-                    multiChain = true;
+                    isMultiChain = true;
                     c++;
                 }
             } else {
@@ -829,7 +823,6 @@ namespace Structure {
                     && chains.offsets[c + 1] - chains.offsets[c] === 1
                     && chains.offsets[c + 2] - chains.offsets[c + 1] === 1
                 ) {
-                    singleAtomResidues = true;
                     const e1 = index.getEntityFromChain(c);
                     const e2 = index.getEntityFromChain(c + 1 as ChainIndex);
                     if (e1 !== e2) break;
@@ -842,26 +835,18 @@ namespace Structure {
                     const op2 = atomicChainOperatorMappinng.get(c + 1 as ChainIndex);
                     if (op1 !== op2) break;
 
-                    multiChain = true;
+                    isMultiChain = true;
                     c++;
                 }
             }
 
             const elements = SortedArray.ofBounds(start as ElementIndex, chains.offsets[c + 1] as ElementIndex);
 
-            if (PARTITION) {
-                // check for polymer to exclude CA/P-only models
-                if (singleAtomResidues && !isPolymerChain(model, c)) {
-                    partitionAtomicUnitByAtom(model, elements, builder, multiChain, operator);
-                } else if (elements.length > 200000 || isWaterChain(model, c)) {
-                    // split up very large chains e.g. lipid bilayers, micelles or water with explicit H
-                    partitionAtomicUnitByResidue(model, elements, builder, multiChain, operator);
-                } else {
-                    builder.addUnit(Unit.Kind.Atomic, model, operator, elements, multiChain ? Unit.Trait.MultiChain : Unit.Trait.None);
-                }
-            } else {
-                builder.addUnit(Unit.Kind.Atomic, model, operator, elements, multiChain ? Unit.Trait.MultiChain : Unit.Trait.None);
-            }
+            let traits = Unit.Trait.None;
+            if (isMultiChain) traits |= Unit.Trait.MultiChain;
+            if (isWater) traits |= Unit.Trait.Water;
+
+            builder.addUnit(Unit.Kind.Atomic, model, operator, elements, traits);
         }
 
         const cs = model.coarseHierarchy;
@@ -880,70 +865,6 @@ namespace Structure {
     function isWaterChain(model: Model, chainIndex: ChainIndex) {
         const e = model.atomicHierarchy.index.getEntityFromChain(chainIndex);
         return model.entities.data.type.value(e) === 'water';
-    }
-
-    function isPolymerChain(model: Model, chainIndex: ChainIndex) {
-        const e = model.atomicHierarchy.index.getEntityFromChain(chainIndex);
-        return model.entities.data.type.value(e) === 'polymer';
-    }
-
-    function partitionAtomicUnitByAtom(model: Model, indices: SortedArray, builder: StructureBuilder, multiChain: boolean, operator: SymmetryOperator) {
-        const { x, y, z } = model.atomicConformation;
-        const position = { x, y, z, indices };
-        const lookup = GridLookup3D(position, getBoundary(position), 8192);
-        const { offset, count, array } = lookup.buckets;
-
-        const traits = (multiChain ? Unit.Trait.MultiChain : Unit.Trait.None) | (offset.length > 1 ? Unit.Trait.Partitioned : Unit.Trait.None);
-
-        builder.beginChainGroup();
-        for (let i = 0, _i = offset.length; i < _i; i++) {
-            const start = offset[i];
-            const set = new Int32Array(count[i]);
-            for (let j = 0, _j = count[i]; j < _j; j++) {
-                set[j] = indices[array[start + j]];
-            }
-            builder.addUnit(Unit.Kind.Atomic, model, operator, SortedArray.ofSortedArray(set), traits);
-        }
-        builder.endChainGroup();
-    }
-
-    // keeps atoms of residues together
-    function partitionAtomicUnitByResidue(model: Model, indices: SortedArray, builder: StructureBuilder, multiChain: boolean, operator: SymmetryOperator) {
-        const { residueAtomSegments } = model.atomicHierarchy;
-
-        const startIndices: number[] = [];
-        const endIndices: number[] = [];
-
-        const residueIt = Segmentation.transientSegments(residueAtomSegments, indices);
-        while (residueIt.hasNext) {
-            const residueSegment = residueIt.move();
-            startIndices[startIndices.length] = indices[residueSegment.start];
-            endIndices[endIndices.length] = indices[residueSegment.end];
-        }
-
-        const firstResidueAtomCount = endIndices[0] - startIndices[0];
-        const gridCellCount = 512 * firstResidueAtomCount;
-
-        const { x, y, z } = model.atomicConformation;
-        const position = { x, y, z, indices: SortedArray.ofSortedArray(startIndices) };
-        const lookup = GridLookup3D(position, getBoundary(position), gridCellCount);
-        const { offset, count, array } = lookup.buckets;
-
-        const traits = (multiChain ? Unit.Trait.MultiChain : Unit.Trait.None) | (offset.length > 1 ? Unit.Trait.Partitioned : Unit.Trait.None);
-
-        builder.beginChainGroup();
-        for (let i = 0, _i = offset.length; i < _i; i++) {
-            const start = offset[i];
-            const set: number[] = [];
-            for (let j = 0, _j = count[i]; j < _j; j++) {
-                const k = array[start + j];
-                for (let l = startIndices[k], _l = endIndices[k]; l < _l; l++) {
-                    set[set.length] = l;
-                }
-            }
-            builder.addUnit(Unit.Kind.Atomic, model, operator, SortedArray.ofSortedArray(new Int32Array(set)), traits);
-        }
-        builder.endChainGroup();
     }
 
     function addCoarseUnits(builder: StructureBuilder, model: Model, elements: CoarseElements, kind: Unit.Kind) {